requirements.txt
Pillow==8.4.0
requests
instabot